module 3. task 5. fix 2

+++ src/main/java/lang/print/gaps/task5/NumberDivider.java
@@ -1,7 +1,7 @@
 package lang.print.gaps.task5;
 
 public class NumberDivider {
-    void divide(int toBeDivided) {
+    void divide(double toBeDivided) {
         double res = toBeDivided / 5;
         System.out.println(res);
     }
